Updated Kotlin and other packages versions.
* In addproblem/generic/SsAddGenericProblemViewModel, created a data class object that can be used by each question, and as a result, be organized better.

In addition, this class allows it so that the class can be passed into an object and one of its attributes can be modified, namely, the index. This makes index initialization for each question much more flexible.

* In addproblem/generic/question/Ss*Question, access the Question data class mentioned above to access the index and question of the respective question.

When the index is accessed, if the index for this question is not set yet, it is set and saved.

* In db/generic/SsGenericProblem, added "observableName".

* In onboarding/SsAskAboutClimbsPage, added an icon next to the climb name, and changed the CheckBox (on the left) to a Switch (on the right).

* In onboarding/SsAskAboutGradingSystemsPage, increased the padding to space out each climb type.

* In ui/SsButton, changed the "(un)checkedColor" function arguments to instead of anonymous functions that are passed the current state and return the desired type. This way the logic can be contained there and configured by the user, if they want, instead of being fixed in the function itself.

These new arguments are "colors" and "border".

Also, this changes comes along with changing the default color function to animate the color, instead of sharply transitioning from one color to another.